Move imports to inline component
